Refactor contribution permissions to use centralized permission class
