+++ contributions/permissions.py
@@ -1,5 +1,7 @@
 from rest_framework import permissions
-from chama.models import Chama
+from rest_framework.exceptions import PermissionDenied
+from chama.models import Chama, Membership
+from .models import Contribution
 
 class IsChamaMember(permissions.BasePermission):
     """Only members of the chama can create / view contributions."""
@@ -18,3 +20,25 @@ class IsChamaMember(permissions.BasePermission):
             return False
             
         return self._is_member(request.user, chama_id)
+
+    def _check_permissions(self, contribution):
+        # Check if user has permission to update/delete a contribution
+        try:
+            membership = Membership.objects.get(user=self.request.user, chama=contribution.chama)
+        except Membership.DoesNotExist:
+            raise PermissionDenied("You are not a member of this Chama.")
+        
+        if membership.status != Membership.Status.ACTIVE:
+            raise PermissionDenied("Only users with ACTIVE membership can perform this action.")
+        
+        return membership.role == Membership.Role.ADMIN, membership
+    
+    def partial_update_permission(self, request, contribution):
+        # Memebrs can only delete their own PENDING or FAILED contributions
+        if contribution.member.user != request.user:
+            raise PermissionDenied("You can only delete your own contributions.")
+        
+        if contribution.status not in [Contribution.Status.PENDING, Contribution.Status.FAILED]:
+            raise PermissionDenied("Only pending or rejected contributions can be deleted.")
+        
+        return True
